consolidate custom column e2e specs
