+++ e2e/test/scenarios/custom-column/custom-column.cy.spec.js
@@ -20,9 +20,11 @@ import {
   tableHeaderClick,
   visitQuestionAdhoc,
   visualize,
+  openProductsTable,
+  openTable,
 } from "e2e/support/helpers";
 
-const { ORDERS, ORDERS_ID, PRODUCTS, PRODUCTS_ID } = SAMPLE_DATABASE;
+const { ORDERS, ORDERS_ID, PRODUCTS, PRODUCTS_ID, PEOPLE_ID } = SAMPLE_DATABASE;
 
 describe("scenarios > question > custom column", () => {
   beforeEach(() => {
@@ -723,5 +725,299 @@ describe("scenarios > question > custom column", () => {
 
       checkExpressionEditorHelperPopoverPosition();
     });
+  });
+});
+
+describe("scenarios > question > custom column > data type", () => {
+  function addCustomColumns(columns) {
+    cy.wrap(columns).each((column, index) => {
+      if (index) {
+        getNotebookStep("expression").icon("add").click();
+      } else {
+        cy.findByLabelText("Custom column").click();
+      }
+
+      enterCustomColumnDetails(column);
+      cy.button("Done").click({ force: true });
+    });
+  }
+
+  function openCustomColumnInTable(table) {
+    openTable({ table, mode: "notebook" });
+    cy.findByText("Custom column").click();
+  }
+
+  beforeEach(() => {
+    restore();
+    restore("postgres-12");
+
+    cy.signInAsAdmin();
+  });
+
+  it("should understand string functions (metabase#13217)", () => {
+    openCustomColumnInTable(PRODUCTS_ID);
+
+    enterCustomColumnDetails({
+      formula: "concat([Category], [Title])",
+      name: "CategoryTitle",
+    });
+
+    cy.button("Done").click();
+
+    filter({ mode: "notebook" });
+
+    popover().within(() => {
+      cy.findByText("CategoryTitle").click();
+      cy.findByPlaceholderText("Enter a number").should("not.exist");
+      cy.findByPlaceholderText("Enter some text").should("be.visible");
+    });
+  });
+
+  it("should understand date functions", () => {
+    startNewQuestion();
+    entityPickerModal().within(() => {
+      entityPickerModalTab("Tables").click();
+      cy.findByText("QA Postgres12").click();
+      cy.findByText("Orders").click();
+    });
+
+    addCustomColumns([
+      { name: "Year", formula: "year([Created At])" },
+      { name: "Quarter", formula: "quarter([Created At])" },
+      { name: "Month", formula: "month([Created At])" },
+      { name: "Week", formula: 'week([Created At], "iso")' },
+      { name: "Day", formula: "day([Created At])" },
+      { name: "Weekday", formula: "weekday([Created At])" },
+      { name: "Hour", formula: "hour([Created At])" },
+      { name: "Minute", formula: "minute([Created At])" },
+      { name: "Second", formula: "second([Created At])" },
+      {
+        name: "Datetime Add",
+        formula: 'datetimeAdd([Created At], 1, "month")',
+      },
+      {
+        name: "Datetime Subtract",
+        formula: 'datetimeSubtract([Created At], 1, "month")',
+      },
+      {
+        name: "ConvertTimezone 3 args",
+        formula: 'convertTimezone([Created At], "Asia/Ho_Chi_Minh", "UTC")',
+      },
+      {
+        name: "ConvertTimezone 2 args",
+        formula: 'convertTimezone([Created At], "Asia/Ho_Chi_Minh")',
+      },
+    ]);
+
+    visualize();
+  });
+
+  it("should relay the type of a date field", () => {
+    openCustomColumnInTable(PEOPLE_ID);
+
+    enterCustomColumnDetails({ formula: "[Birth Date]", name: "DoB" });
+    cy.button("Done").click();
+
+    filter({ mode: "notebook" });
+    popover().within(() => {
+      cy.findByText("DoB").click();
+      cy.findByPlaceholderText("Enter a number").should("not.exist");
+      cy.findByText("Relative dates…").click();
+      cy.findByText("Past").click();
+      cy.findByDisplayValue("days").should("be.visible");
+    });
+  });
+
+  it("should handle CASE (metabase#13122)", () => {
+    openCustomColumnInTable(ORDERS_ID);
+
+    enterCustomColumnDetails({
+      formula: "case([Discount] > 0, [Created At], [Product → Created At])",
+      name: "MiscDate",
+    });
+    cy.button("Done").click();
+
+    filter({ mode: "notebook" });
+    popover().within(() => {
+      cy.findByText("MiscDate").click();
+      cy.findByPlaceholderText("Enter a number").should("not.exist");
+
+      cy.findByText("Relative dates…").click();
+      cy.findByText("Past").click();
+      cy.findByDisplayValue("days").should("be.visible");
+    });
+  });
+
+  it("should handle COALESCE", () => {
+    openCustomColumnInTable(ORDERS_ID);
+
+    enterCustomColumnDetails({
+      formula: "COALESCE([Product → Created At], [Created At])",
+      name: "MiscDate",
+    });
+    cy.button("Done").click();
+
+    filter({ mode: "notebook" });
+    popover().within(() => {
+      cy.findByText("MiscDate").click();
+      cy.findByPlaceholderText("Enter a number").should("not.exist");
+      cy.findByText("Relative dates…").click();
+      cy.findByText("Past").click();
+      cy.findByDisplayValue("days").should("be.visible");
+    });
+  });
+});
+
+describe("scenarios > question > custom column > error feedback", () => {
+  beforeEach(() => {
+    restore();
+    cy.signInAsAdmin();
+
+    openProductsTable({ mode: "notebook" });
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.findByText("Custom column").click();
+  });
+
+  it("should catch non-existent field reference", () => {
+    enterCustomColumnDetails({
+      formula: "abcdef",
+      name: "Non-existent",
+    });
+
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.contains(/^Unknown Field: abcdef/i);
+  });
+
+  it("should fail on expression validation errors", () => {
+    enterCustomColumnDetails({
+      formula: "SUBSTRING('foo', 0, 1)",
+      name: "BadSubstring",
+    });
+
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.contains(/positive integer/i);
+  });
+});
+
+// ExpressionEditorTextfield jsx component
+describe("scenarios > question > custom column > expression editor", () => {
+  beforeEach(() => {
+    restore();
+    cy.signInAsAdmin();
+
+    // This is the default screen size but we need it explicitly set for this test because of the resize later on
+    cy.viewport(1280, 800);
+
+    openOrdersTable({ mode: "notebook" });
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.findByText("Custom column").click();
+
+    enterCustomColumnDetails({
+      formula: "1+1", // Formula was intentionally written without spaces (important for this repro)!
+      name: "Math",
+    });
+    cy.button("Done").should("not.be.disabled");
+  });
+
+  /**
+   * We abuse {force: true} arguments below because AceEditor cannot be found
+   * on a second click and type commands (the first ones happen in the beforeEach block above )
+   */
+  it("should not accidentally delete Custom Column formula value and/or Custom Column name (metabase#15734)", () => {
+    cy.get("@formula")
+      .click({ force: true })
+      .type("{movetoend}{leftarrow}{movetostart}{rightarrow}{rightarrow}", {
+        force: true,
+      });
+    cy.findByDisplayValue("Math").focus();
+    cy.button("Done").should("not.be.disabled");
+  });
+
+  /**
+   * 1. Explanation for `cy.get("@formula").click();`
+   *  - Without it, test runner is too fast and the test results in false positive.
+   *  - This gives it enough time to update the DOM. The same result can be achieved with `cy.wait(1)`
+   */
+  it("should not erase Custom column formula and Custom column name when expression is incomplete (metabase#16126)", () => {
+    cy.get("@formula")
+      .focus()
+      .click({ force: true })
+      .type("{movetoend}{backspace}", { force: true })
+      .blur();
+
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.findByText("Expected expression");
+    cy.button("Done").should("be.disabled");
+  });
+
+  it("should not erase Custom Column formula and Custom Column name on window resize (metabase#16127)", () => {
+    cy.viewport(1260, 800);
+    cy.findByDisplayValue("Math");
+    cy.button("Done").should("not.be.disabled");
+  });
+});
+
+describe("scenarios > question > custom column > help text", () => {
+  beforeEach(() => {
+    restore();
+    cy.signInAsAdmin();
+
+    openProductsTable({ mode: "notebook" });
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.findByText("Custom column").click();
+  });
+
+  it("should appear while inside a function", () => {
+    enterCustomColumnDetails({ formula: "Lower(", blur: false });
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.contains("lower(text)");
+  });
+
+  it("should appear after a field reference", () => {
+    enterCustomColumnDetails({ formula: "Lower([Category]", blur: false });
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.contains("lower(text)");
+  });
+
+  it("should not appear while outside a function", () => {
+    enterCustomColumnDetails({ formula: "Lower([Category])", blur: false });
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.findByText("lower(text)").should("not.exist");
+  });
+
+  it("should not appear when formula field is not in focus (metabase#15891)", () => {
+    enterCustomColumnDetails({
+      formula: "rou{enter}1.5){leftArrow}",
+      blur: false,
+    });
+
+    cy.findByTestId("expression-helper-popover").findByText(
+      "round([Temperature])",
+    );
+
+    cy.log("Blur event should remove the expression helper popover");
+    cy.get("@formula").blur();
+    cy.findByTestId("expression-helper-popover").should("not.exist");
+
+    cy.get("@formula").focus();
+    cy.findByTestId("expression-helper-popover").findByText(
+      "round([Temperature])",
+    );
+
+    cy.log(
+      "Pressing `escape` key should also remove the expression helper popover",
+    );
+    cy.get("@formula").type("{esc}");
+    cy.findByTestId("expression-helper-popover").should("not.exist");
+  });
+
+  it("should not disappear when clicked on (metabase#17548)", () => {
+    enterCustomColumnDetails({ formula: "rou{enter}", blur: false });
+
+    // Shouldn't hide on click
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.findByText("round([Temperature])").click();
+    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
+    cy.findByText("round([Temperature])");
   });
 });
