--- e2e/test/scenarios/custom-column/reproductions/18814-cc-used-in-aggregation-for-nested-query.cy.spec.js
@@ -1,45 +1,0 @@
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { restore, popover, visualize } from "e2e/support/helpers";
-
-const { ORDERS_ID, ORDERS } = SAMPLE_DATABASE;
-
-const ccName = "Custom Created At";
-
-const questionDetails = {
-  name: "18814",
-  query: {
-    "source-query": {
-      "source-table": ORDERS_ID,
-      aggregation: [["count"]],
-      breakout: [["field", ORDERS.CREATED_AT, { "temporal-unit": "year" }]],
-    },
-    expressions: {
-      [ccName]: ["field", "CREATED_AT", { "base-type": "type/DateTime" }],
-    },
-  },
-};
-
-describe("issue 18814", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-
-    cy.createQuestion(questionDetails, { visitQuestion: true });
-  });
-
-  it("should be able to use a custom column in aggregation for a nested query (metabase#18814)", () => {
-    cy.icon("notebook").click();
-
-    cy.icon("sum").click();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Count of rows").click();
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Pick a column to group by").click();
-    popover().contains(ccName).click();
-
-    visualize();
-
-    cy.findByTestId("query-visualization-root").should("contain", "2022");
-  });
-});
